requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
Pillow==10.2.0
google-generativeai==0.3.2
python-dateutil==2.8.2
requests==2.31.0
customtkinter==5.2.2
pystray==0.19.5
pyinstaller==6.5.0 